Refactor access control in ordersController.mjs to handle admin and user permissions

+++ server/controllers/ordersController.mjs
@@ -4,7 +4,15 @@ export const getOrderById = async (req, res) => {
   try {
     const order = await Order.findById(req.params.id).populate("user");
     if (order) {
-      res.json(order);
+      if (
+        req.user.isAdmin ||
+        order.user.toString() === req.user._id.toString()
+      ) {
+        res.json(order);
+      } else {
+        res.status(403);
+        throw new Error("Unauthorized access");
+      }
     } else {
       res.status(404);
       throw new Error("Order not found");
@@ -22,21 +30,40 @@ export const getAllOrders = async (req, res) => {
     const startIndex = (page - 1) * limit;
     const user = req.user;
 
-    const totalOrders = await Order.countDocuments({ user: user._id });
-    const totalPages = Math.ceil(totalOrders / limit);
+    // Check if the user is an admin
+    if (user.isAdmin) {
+      const totalOrders = await Order.countDocuments();
+      const totalPages = Math.ceil(totalOrders / limit);
 
-    const orders = await Order.find({ user: user._id })
-      .populate("user")
-      .skip(startIndex)
-      .limit(limit);
+      const orders = await Order.find()
+        .populate("user")
+        .skip(startIndex)
+        .limit(limit);
 
-    const pagination = {
-      currentPage: page,
-      totalPages: totalPages,
-      totalOrders: totalOrders,
-    };
+      const pagination = {
+        currentPage: page,
+        totalPages: totalPages,
+        totalOrders: totalOrders,
+      };
 
-    res.json({ orders, pagination });
+      res.json({ orders, pagination });
+    } else {
+      const totalOrders = await Order.countDocuments({ user: user._id });
+      const totalPages = Math.ceil(totalOrders / limit);
+
+      const orders = await Order.find({ user: user._id })
+        .populate("user")
+        .skip(startIndex)
+        .limit(limit);
+
+      const pagination = {
+        currentPage: page,
+        totalPages: totalPages,
+        totalOrders: totalOrders,
+      };
+
+      res.json({ orders, pagination });
+    }
   } catch (error) {
     res.status(500);
     throw new Error("Error while fetching orders");
@@ -84,6 +111,12 @@ export const updateOrderToAccepted = async (req, res) => {
     const order = await Order.findById(req.params.id);
 
     if (order) {
+      // Check if the user is an admin
+      if (!req.user.isAdmin) {
+        res.status(403);
+        throw new Error("Unauthorized access");
+      }
+
       order.status = "accepted";
 
       const updatedOrder = await order.save();
@@ -103,6 +136,12 @@ export const updateOrderToRejected = async (req, res) => {
     const order = await Order.findById(req.params.id);
 
     if (order) {
+      // Check if the user is an admin
+      if (!req.user.isAdmin) {
+        res.status(403);
+        throw new Error("Unauthorized access");
+      }
+
       order.status = "rejected";
 
       const updatedOrder = await order.save();
@@ -121,12 +160,21 @@ export const deleteOrder = async (req, res) => {
   try {
     const order = await Order.findById(req.params.id);
     if (order) {
-      if (order.status === "pending") {
-        await order.deleteOne();
-        res.json({ message: "Order deleted" });
+      // Check if the user is an admin or the order belongs to the user
+      if (
+        req.user.isAdmin ||
+        order.user.toString() === req.user._id.toString()
+      ) {
+        if (order.status === "pending") {
+          await order.deleteOne();
+          res.json({ message: "Order deleted" });
+        } else {
+          res.status(400);
+          throw new Error("Order can only be cancelled if it is pending");
+        }
       } else {
-        res.status(400);
-        throw new Error("Order only be cancelled if it is pending");
+        res.status(403);
+        throw new Error("Unauthorized access");
       }
     } else {
       res.status(404);
